There is a problem with the creation of the zero fill and no-data filled to 1 rasters. Both are being filled to all zeros and all ones, respectively. This leads to very odd CPG creation.
The current implementation pulls these rasters into rasterio, pull the no data value, and uses that to decide where to fill in zeros or ones. For the no-data filled to 1, the raster is copied to zeros, filled totally to ones, then the cells that do not contain the no-data value are turned to zero.